Add api/filter_test
