+++ waku/v2/protocol/filter/filter_test.go
@@ -2,25 +2,16 @@ package filter
 
 import (
 	"context"
-	"crypto/rand"
 	"errors"
-	"fmt"
-	"strconv"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/libp2p/go-libp2p/core/host"
-	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-libp2p/core/peerstore"
-	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/suite"
 	"github.com/waku-org/go-waku/tests"
 	"github.com/waku-org/go-waku/waku/v2/protocol"
-	"github.com/waku-org/go-waku/waku/v2/protocol/relay"
-	"github.com/waku-org/go-waku/waku/v2/protocol/subscription"
 	"github.com/waku-org/go-waku/waku/v2/service"
-	"github.com/waku-org/go-waku/waku/v2/timesource"
 	"github.com/waku-org/go-waku/waku/v2/utils"
 	"go.uber.org/zap"
 )
@@ -32,309 +23,9 @@ func TestFilterSuite(t *testing.T) {
 const defaultTestPubSubTopic = "/waku/2/go/filter/test"
 const defaultTestContentTopic = "/test/10/my-app"
 
-type FilterTestSuite struct {
-	suite.Suite
-
-	testTopic        string
-	testContentTopic string
-	ctx              context.Context
-	ctxCancel        context.CancelFunc
-	lightNode        *WakuFilterLightNode
-	lightNodeHost    host.Host
-	relayNode        *relay.WakuRelay
-	relaySub         *relay.Subscription
-	fullNode         *WakuFilterFullNode
-	fullNodeHost     host.Host
-	wg               *sync.WaitGroup
-	contentFilter    protocol.ContentFilter
-	subDetails       []*subscription.SubscriptionDetails
-	log              *zap.Logger
-}
-
-type WakuMsg struct {
-	pubSubTopic  string
-	contentTopic string
-	payload      string
-}
-
-func (s *FilterTestSuite) makeWakuRelay(topic string, shared bool) (*relay.WakuRelay, *relay.Subscription, host.Host, relay.Broadcaster) {
-
-	broadcaster := relay.NewBroadcaster(10)
-	s.Require().NoError(broadcaster.Start(context.Background()))
-
-	port, err := tests.FindFreePort(s.T(), "", 5)
-	s.Require().NoError(err)
-
-	host, err := tests.MakeHost(context.Background(), port, rand.Reader)
-	s.Require().NoError(err)
-
-	relay := relay.NewWakuRelay(broadcaster, 0, timesource.NewDefaultClock(), prometheus.DefaultRegisterer, s.log)
-	relay.SetHost(host)
-
-	if shared {
-		s.fullNodeHost = host
-	}
-
-	err = relay.Start(context.Background())
-	s.Require().NoError(err)
-
-	sub, err := relay.Subscribe(context.Background(), protocol.NewContentFilter(topic))
-	s.Require().NoError(err)
-
-	return relay, sub[0], host, broadcaster
-}
-
-func (s *FilterTestSuite) makeWakuFilterLightNode(start bool, withBroadcaster bool) *WakuFilterLightNode {
-	port, err := tests.FindFreePort(s.T(), "", 5)
-	s.Require().NoError(err)
-
-	host, err := tests.MakeHost(context.Background(), port, rand.Reader)
-	s.Require().NoError(err)
-	var b relay.Broadcaster
-	if withBroadcaster {
-		b = relay.NewBroadcaster(10)
-		s.Require().NoError(b.Start(context.Background()))
-	}
-	filterPush := NewWakuFilterLightNode(b, nil, timesource.NewDefaultClock(), prometheus.DefaultRegisterer, s.log)
-	filterPush.SetHost(host)
-	s.lightNodeHost = host
-	if start {
-		err = filterPush.Start(context.Background())
-		s.Require().NoError(err)
-	}
-
-	return filterPush
-}
-
-func (s *FilterTestSuite) makeWakuFilterFullNode(topic string, withRegisterAll bool, shared bool) (*relay.WakuRelay, *WakuFilterFullNode) {
-	var sub *relay.Subscription
-	node, relaySub, host, broadcaster := s.makeWakuRelay(topic, shared)
-
-	if shared {
-		s.relaySub = relaySub
-	}
-
-	node2Filter := NewWakuFilterFullNode(timesource.NewDefaultClock(), prometheus.DefaultRegisterer, s.log)
-	node2Filter.SetHost(host)
-
-	if withRegisterAll {
-		sub = broadcaster.RegisterForAll()
-	} else {
-		sub = broadcaster.Register(protocol.NewContentFilter(topic))
-	}
-
-	err := node2Filter.Start(s.ctx, sub)
-	s.Require().NoError(err)
-
-	return node, node2Filter
-}
-
-func (s *FilterTestSuite) waitForMsg(fn func(), ch chan *protocol.Envelope) {
-	s.wg.Add(1)
-	var msgFound = false
-	go func() {
-		defer s.wg.Done()
-		select {
-		case env := <-ch:
-			for _, topic := range s.contentFilter.ContentTopicsList() {
-				if topic == env.Message().GetContentTopic() {
-					msgFound = true
-				}
-			}
-			s.Require().True(msgFound)
-		case <-time.After(1 * time.Second):
-			s.Require().Fail("Message timeout")
-		case <-s.ctx.Done():
-			s.Require().Fail("test exceeded allocated time")
-		}
-	}()
-
-	if fn != nil {
-		fn()
-	}
-
-	s.wg.Wait()
-}
-
-func matchOneOfManyMsg(one WakuMsg, many []WakuMsg) bool {
-	for _, m := range many {
-		if m.pubSubTopic == one.pubSubTopic &&
-			m.contentTopic == one.contentTopic &&
-			m.payload == one.payload {
-			return true
-		}
-	}
-
-	return false
-}
-
-func (s *FilterTestSuite) waitForMessages(fn func(), subs []*subscription.SubscriptionDetails, expected []WakuMsg) {
-	s.wg.Add(1)
-	msgCount := len(expected)
-	found := 0
-	s.log.Info("Expected messages ", zap.String("count", strconv.Itoa(msgCount)))
-	s.log.Info("Existing subscriptions ", zap.String("count", strconv.Itoa(len(subs))))
-
-	go func() {
-		defer s.wg.Done()
-		for _, sub := range subs {
-			s.log.Info("Looking at ", zap.String("pubSubTopic", sub.ContentFilter.PubsubTopic))
-			for i := 0; i < msgCount; i++ {
-				select {
-				case env, ok := <-sub.C:
-					if !ok {
-						continue
-					}
-					received := WakuMsg{
-						pubSubTopic:  env.PubsubTopic(),
-						contentTopic: env.Message().GetContentTopic(),
-						payload:      string(env.Message().GetPayload()),
-					}
-					s.log.Debug("received message ", zap.String("pubSubTopic", received.pubSubTopic), zap.String("contentTopic", received.contentTopic), zap.String("payload", received.payload))
-					if matchOneOfManyMsg(received, expected) {
-						found++
-					}
-				case <-time.After(3 * time.Second):
-
-				case <-s.ctx.Done():
-					s.Require().Fail("test exceeded allocated time")
-				}
-			}
-		}
-	}()
-
-	if fn != nil {
-		fn()
-	}
-
-	s.wg.Wait()
-	s.Require().True(msgCount == found)
-}
-
-func (s *FilterTestSuite) waitForTimeout(fn func(), ch chan *protocol.Envelope) {
-	s.wg.Add(1)
-	go func() {
-		defer s.wg.Done()
-		select {
-		case env, ok := <-ch:
-			if ok {
-				s.Require().Fail("should not receive another message", zap.String("payload", string(env.Message().Payload)))
-			}
-		case <-time.After(1 * time.Second):
-			// Timeout elapsed, all good
-		case <-s.ctx.Done():
-			s.Require().Fail("waitForTimeout test exceeded allocated time")
-		}
-	}()
-
-	fn()
-
-	s.wg.Wait()
-}
-
-func (s *FilterTestSuite) subscribe(pubsubTopic string, contentTopic string, peer peer.ID) []*subscription.SubscriptionDetails {
-
-	for _, sub := range s.subDetails {
-		if sub.ContentFilter.PubsubTopic == pubsubTopic {
-			sub.Add(contentTopic)
-			s.contentFilter = sub.ContentFilter
-			subDetails, err := s.lightNode.Subscribe(s.ctx, s.contentFilter, WithPeer(peer))
-			s.Require().NoError(err)
-			return subDetails
-		}
-	}
-
-	s.contentFilter = protocol.ContentFilter{PubsubTopic: pubsubTopic, ContentTopics: protocol.NewContentTopicSet(contentTopic)}
-
-	subDetails, err := s.lightNode.Subscribe(s.ctx, s.contentFilter, WithPeer(peer))
-	s.Require().NoError(err)
-
-	// Sleep to make sure the filter is subscribed
-	time.Sleep(1 * time.Second)
-
-	return subDetails
-}
-
-func (s *FilterTestSuite) unsubscribe(pubsubTopic string, contentTopic string, peer peer.ID) []*subscription.SubscriptionDetails {
-
-	for _, sub := range s.subDetails {
-		if sub.ContentFilter.PubsubTopic == pubsubTopic {
-			topicsCount := len(sub.ContentFilter.ContentTopicsList())
-			if topicsCount == 1 {
-				_, err := s.lightNode.Unsubscribe(s.ctx, sub.ContentFilter, WithPeer(peer))
-				s.Require().NoError(err)
-			} else {
-				sub.Remove(contentTopic)
-			}
-			s.contentFilter = sub.ContentFilter
-		}
-	}
-
-	return s.lightNode.Subscriptions()
-}
-
-func (s *FilterTestSuite) publishMsg(topic, contentTopic string, optionalPayload ...string) {
-	var payload string
-	if len(optionalPayload) > 0 {
-		payload = optionalPayload[0]
-	} else {
-		payload = "123"
-	}
-
-	_, err := s.relayNode.Publish(s.ctx, tests.CreateWakuMessage(contentTopic, utils.GetUnixEpoch(), payload), relay.WithPubSubTopic(topic))
-	s.Require().NoError(err)
-}
-
-func (s *FilterTestSuite) publishMessages(msgs []WakuMsg) {
-	for _, m := range msgs {
-		_, err := s.relayNode.Publish(s.ctx, tests.CreateWakuMessage(m.contentTopic, utils.GetUnixEpoch(), m.payload), relay.WithPubSubTopic(m.pubSubTopic))
-		s.Require().NoError(err)
-	}
-}
-
-func prepareData(quantity int, topics, contentTopics, payloads bool, sg tests.StringGenerator) []WakuMsg {
-	var (
-		pubsubTopic     = defaultTestPubSubTopic  // Has to be the same with initial s.testTopic
-		contentTopic    = defaultTestContentTopic // Has to be the same with initial s.testContentTopic
-		payload         = "test_msg"
-		messages        []WakuMsg
-		strMaxLenght    = 4097
-		generatedString = ""
-	)
-
-	for i := 0; i < quantity; i++ {
-		msg := WakuMsg{
-			pubSubTopic:  pubsubTopic,
-			contentTopic: contentTopic,
-			payload:      payload,
-		}
-
-		if sg != nil {
-			generatedString, _ = sg(strMaxLenght)
-
-		}
-
-		if topics {
-			msg.pubSubTopic = fmt.Sprintf("%s%02d%s", pubsubTopic, i, generatedString)
-		}
-
-		if contentTopics {
-			msg.contentTopic = fmt.Sprintf("%s%02d%s", contentTopic, i, generatedString)
-		}
-
-		if payloads {
-			msg.payload = fmt.Sprintf("%s%02d%s", payload, i, generatedString)
-		}
-
-		messages = append(messages, msg)
-	}
-
-	return messages
-}
-
 func (s *FilterTestSuite) SetupTest() {
 	log := utils.Logger() //.Named("filterv2-test")
-	s.log = log
+	s.Log = log
 	// Use a pointer to WaitGroup so that to avoid copying
 	// https://pkg.go.dev/sync#WaitGroup
 	s.wg = &sync.WaitGroup{}
@@ -347,11 +38,11 @@ func (s *FilterTestSuite) SetupTest() {
 	s.testTopic = defaultTestPubSubTopic
 	s.testContentTopic = defaultTestContentTopic
 
-	s.lightNode = s.makeWakuFilterLightNode(true, true)
+	s.lightNode = s.StartNode(s.MakeWakuFilterLightNode())
 
 	//TODO: Add tests to verify broadcaster.
 
-	s.relayNode, s.fullNode = s.makeWakuFilterFullNode(s.testTopic, false, true)
+	s.relayNode, s.fullNode = s.MakeWakuFilterFullNode(s.testTopic, false)
 
 	// Connect nodes
 	s.lightNodeHost.Peerstore().AddAddr(s.fullNodeHost.ID(), tests.GetHostAddress(s.fullNodeHost), peerstore.PermanentAddrTTL)
@@ -363,7 +54,7 @@ func (s *FilterTestSuite) SetupTest() {
 func (s *FilterTestSuite) TearDownTest() {
 	s.fullNode.Stop()
 	s.relayNode.Stop()
-	s.relaySub.Unsubscribe()
+	s.RelaySub.Unsubscribe()
 	s.lightNode.Stop()
 	s.ctxCancel()
 }
@@ -410,7 +101,7 @@ func (s *FilterTestSuite) TestStartStop() {
 
 	var wg sync.WaitGroup
 	wg.Add(2)
-	s.lightNode = s.makeWakuFilterLightNode(false, false)
+	s.lightNode = s.MakeWakuFilterLightNode()
 
 	stopNode := func() {
 		for i := 0; i < 100000; i++ {
@@ -454,25 +145,19 @@ func (s *FilterTestSuite) TestAutoShard() {
 	s.testContentTopic = cTopic1Str
 	s.testTopic = pubSubTopic.String()
 
-	s.lightNode = s.makeWakuFilterLightNode(true, false)
-	s.relayNode, s.fullNode = s.makeWakuFilterFullNode(pubSubTopic.String(), false, true)
+	s.lightNode = s.StartNode(s.MakeWakuFilterLightNode())
+	s.relayNode, s.fullNode = s.MakeWakuFilterFullNode(pubSubTopic.String(), false)
 
 	s.lightNodeHost.Peerstore().AddAddr(s.fullNodeHost.ID(), tests.GetHostAddress(s.fullNodeHost), peerstore.PermanentAddrTTL)
 	err = s.lightNodeHost.Peerstore().AddProtocols(s.fullNodeHost.ID(), FilterSubscribeID_v20beta1)
 	s.Require().NoError(err)
 
-	s.log.Info("Testing Autoshard:CreateSubscription")
+	s.Log.Info("Testing Autoshard:CreateSubscription")
 	s.subDetails = s.subscribe("", s.testContentTopic, s.fullNodeHost.ID())
-	s.waitForMsg(func() {
-		_, err := s.relayNode.Publish(s.ctx, tests.CreateWakuMessage(s.testContentTopic, utils.GetUnixEpoch()), relay.WithPubSubTopic(s.testTopic))
-		s.Require().NoError(err)
-
-	}, s.subDetails[0].C)
+	s.waitForMsg(&WakuMsg{s.testTopic, s.testContentTopic, ""}, s.subDetails[0].C)
 
 	// Wrong content topic
-	s.waitForTimeout(func() {
-		s.publishMsg(s.testTopic, "TopicB", "second")
-	}, s.subDetails[0].C)
+	s.waitForTimeout(&WakuMsg{s.testTopic, "TopicB", "second"}, s.subDetails[0].C)
 
 	_, err = s.lightNode.Unsubscribe(s.ctx, s.contentFilter, WithPeer(s.fullNodeHost.ID()))
 	s.Require().NoError(err)
@@ -480,27 +165,21 @@ func (s *FilterTestSuite) TestAutoShard() {
 	time.Sleep(1 * time.Second)
 
 	// Should not receive after unsubscribe
-	s.waitForTimeout(func() {
-		s.publishMsg(s.testTopic, s.testContentTopic, "third")
-	}, s.subDetails[0].C)
+	s.waitForTimeout(&WakuMsg{s.testTopic, s.testContentTopic, "third"}, s.subDetails[0].C)
 
 	s.subDetails = s.subscribe("", s.testContentTopic, s.fullNodeHost.ID())
 
-	s.log.Info("Testing Autoshard:SubscriptionPing")
+	s.Log.Info("Testing Autoshard:SubscriptionPing")
 	err = s.lightNode.Ping(context.Background(), s.fullNodeHost.ID())
 	s.Require().NoError(err)
 
 	// Test ModifySubscription Subscribe to another content_topic
-	s.log.Info("Testing Autoshard:ModifySubscription")
+	s.Log.Info("Testing Autoshard:ModifySubscription")
 
 	newContentTopic := "0/test/1/testTopic1/proto"
 	s.subDetails = s.subscribe("", newContentTopic, s.fullNodeHost.ID())
 
-	s.waitForMsg(func() {
-		_, err := s.relayNode.Publish(s.ctx, tests.CreateWakuMessage(newContentTopic, utils.GetUnixEpoch()), relay.WithPubSubTopic(s.testTopic))
-		s.Require().NoError(err)
-
-	}, s.subDetails[0].C)
+	s.waitForMsg(&WakuMsg{s.testTopic, newContentTopic, ""}, s.subDetails[0].C)
 
 	_, err = s.lightNode.Unsubscribe(s.ctx, protocol.ContentFilter{
 		PubsubTopic:   s.testTopic,
@@ -537,9 +216,9 @@ func (s *FilterTestSuite) TestLightNodeIsListening() {
 }
 
 func (s *FilterTestSuite) BeforeTest(suiteName, testName string) {
-	s.log.Info("Executing ", zap.String("testName", testName))
+	s.Log.Info("Executing ", zap.String("testName", testName))
 }
 
 func (s *FilterTestSuite) AfterTest(suiteName, testName string) {
-	s.log.Info("Finished executing ", zap.String("testName", testName))
+	s.Log.Info("Finished executing ", zap.String("testName", testName))
 }
